Insert doneish, refactor
