+++ src/psqlwrapper/translator/translator.rs
@@ -1,0 +1,43 @@
+use crate::ast::*;
+
+
+
+pub fn translate (stmt: &Stmt) -> String {
+    match stmt {
+        Stmt::Insert(ins) => super::insert::translate(ins),
+        _ => unimplemented!(),
+    }
+}
+
+
+
+
+
+
+#[cfg(test)]
+pub mod test {
+    use super::*;
+    use crate::grammar::StmtParser;
+    #[test]
+    pub fn test_translate_pattern(){
+        let parser = StmtParser::new();
+        let input = vec![
+            r#"INSERT INTO table (col1,col2) VALUES (5,8);"#,
+            r#"INSERT INTO table (col1) VALUES (Val1(5, 2));"#,
+        ];
+        let output = vec![
+            r#"INSERT INTO table (col1,col2) VALUES (5,8);"#,
+            r#"INSERT INTO table (col1) VALUES ('{"Val1":[5,2]}');"#,
+        ];
+        let asts = input.iter().map(|x| parser.parse(x).unwrap()).collect::<Vec<Stmt>>();
+        
+
+        for (ast, out) in asts.iter().zip(output) {
+
+            assert_eq!(translate(ast), out);
+
+        }
+        
+    }
+}
+
